requirements.txt
django>=4.0,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
pyjwt[crypto]==2.9.0
cryptography==44.0.0
cffi==1.17.1
pycparser==2.22
pandas>=2.0
torch>=1.9.0
transformers>=4.20.0
numpy>=1.21.0
scikit-learn>=1.3.0
datasets>=2.0.0
accelerate>=0.20.0
tabulate>=0.9.0
PyMySQL>=1.0.2
python-dotenv>=1.0.0
reportlab>=4.0.0
matplotlib>=3.5.0